# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TravelOfferCategory = typeof TravelOfferCategory[keyof typeof TravelOfferCategory];


export const TravelOfferCategory = {
  circuit: 'circuit',
  sejour: 'sejour',
  croisiere: 'croisiere',
  'city-break': 'city-break',
} as const;

export interface TravelOffer {
  id: string;
  title: string;
  destination: string;
  country: string;
  duration: string;
  price: number;
  currency: string;
  departures: string[];
  category: TravelOfferCategory;
  highlights: string[];
  description: string;
  rating: number;
  reviews: number;
}

export type OmraPackageType = typeof OmraPackageType[keyof typeof OmraPackageType];


export const OmraPackageType = {
  economique: 'economique',
  standard: 'standard',
  premium: 'premium',
  vip: 'vip',
} as const;

export interface OmraPackage {
  id: string;
  title: string;
  duration: string;
  price: number;
  currency: string;
  departures: string[];
  departure_city: string;
  hotel_mecca: string;
  hotel_medina: string;
  stars_mecca: number;
  stars_medina: number;
  nights_mecca: number;
  nights_medina: number;
  includes: string[];
  description: string;
  type: OmraPackageType;
}

export type QuoteRequestTripType = typeof QuoteRequestTripType[keyof typeof QuoteRequestTripType];


export const QuoteRequestTripType = {
  voyage: 'voyage',
  omra: 'omra',
  hajj: 'hajj',
  'sur-mesure': 'sur-mesure',
} as const;

export interface QuoteRequest {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  tripType: QuoteRequestTripType;
  destination?: string;
  departureCity?: string;
  travelers?: string;
  budget?: string;
  departureDate?: string;
  message?: string;
}

export interface QuoteResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  success: boolean;
  error: string;
}