# lib/api-zod/src/generated/types/omraPackage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OmraPackageType } from './omraPackageType';

export interface OmraPackage {
  id: string;
  title: string;
  duration: string;
  price: number;
  currency: string;
  departures: string[];
  departure_city: string;
  hotel_mecca: string;
  hotel_medina: string;
  stars_mecca: number;
  stars_medina: number;
  nights_mecca: number;
  nights_medina: number;
  includes: string[];
  description: string;
  type: OmraPackageType;
}