# lib/api-zod/src/generated/types/quoteRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteRequestTripType } from './quoteRequestTripType';

export interface QuoteRequest {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  tripType: QuoteRequestTripType;
  destination?: string;
  departureCity?: string;
  travelers?: string;
  budget?: string;
  departureDate?: string;
  message?: string;
}