# lib/api-zod/src/generated/types/travelOffer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TravelOfferCategory } from './travelOfferCategory';

export interface TravelOffer {
  id: string;
  title: string;
  destination: string;
  country: string;
  duration: string;
  price: number;
  currency: string;
  departures: string[];
  category: TravelOfferCategory;
  highlights: string[];
  description: string;
  rating: number;
  reviews: number;
}